Refactoring Tests to isolate to issue and specific ReplayData
