+++ tests/Organization.py
@@ -430,20 +430,6 @@ class Organization(Framework.TestCase):
         self.assertTrue(repo.private)
 
     @mock.patch("github.PublicKey.encrypt")
-    def testCreateActionsSecret(self, encrypt):
-        # encrypt returns a non-deterministic value, we need to mock it so the replay data matches
-        encrypt.return_value = "M+5Fm/BqTfB90h3nC7F3BoZuu3nXs+/KtpXwxm9gG211tbRo0F5UiN0OIfYT83CKcx9oKES9Va4E96/b"
-        secret = self.org.create_secret("secret-name", "secret-value", visibility="all")
-        self.assertIsNotNone(secret)
-
-    @mock.patch("github.PublicKey.encrypt")
-    def testCreateDependabotSecret(self, encrypt):
-        # encrypt returns a non-deterministic value, we need to mock it so the replay data matches
-        encrypt.return_value = "M+5Fm/BqTfB90h3nC7F3BoZuu3nXs+/KtpXwxm9gG211tbRo0F5UiN0OIfYT83CKcx9oKES9Va4E96/b"
-        secret = self.org.create_secret("secret-name", "secret-value", secret_type="dependabot", visibility="all")
-        self.assertIsNotNone(secret)
-
-    @mock.patch("github.PublicKey.encrypt")
     def testCreateSecretSelected(self, encrypt):
         repos = [self.org.get_repo("TestPyGithub"), self.org.get_repo("FatherBeaver")]
         # encrypt returns a non-deterministic value, we need to mock it so the replay data matches
