+++ tests/Repository.py
@@ -482,20 +482,6 @@ class Repository(Framework.TestCase):
         self.assertTrue(without_payload)
 
     @mock.patch("github.PublicKey.encrypt")
-    def testCreateActionsSecret(self, encrypt):
-        # encrypt returns a non-deterministic value, we need to mock it so the replay data matches
-        encrypt.return_value = "M+5Fm/BqTfB90h3nC7F3BoZuu3nXs+/KtpXwxm9gG211tbRo0F5UiN0OIfYT83CKcx9oKES9Va4E96/b"
-        secret = self.repo.create_secret("secret-name", "secret-value", "actions")
-        self.assertIsNotNone(secret)
-
-    @mock.patch("github.PublicKey.encrypt")
-    def testCreateDependabotSecret(self, encrypt):
-        # encrypt returns a non-deterministic value, we need to mock it so the replay data matches
-        encrypt.return_value = "M+5Fm/BqTfB90h3nC7F3BoZuu3nXs+/KtpXwxm9gG211tbRo0F5UiN0OIfYT83CKcx9oKES9Va4E96/b"
-        secret = self.repo.create_secret("secret-name", "secret-value", "dependabot")
-        self.assertIsNotNone(secret)
-
-    @mock.patch("github.PublicKey.encrypt")
     def testRepoSecrets(self, encrypt):
         # encrypt returns a non-deterministic value, we need to mock it so the replay data matches
         encrypt.return_value = "M+5Fm/BqTfB90h3nC7F3BoZuu3nXs+/KtpXwxm9gG211tbRo0F5UiN0OIfYT83CKcx9oKES9Va4E96/b"
